Added component for client sided react components
I have no idea why testing worked locally before and now doesn't all of a sudden on the main branch. This fixes that issue.
